test(core): refactor core2 tests
Fixes #229
