+++ packages/app/src/test/core/command-not-found-suggestions.ts
@@ -21,7 +21,7 @@ import * as common from '@kui/core/tests/lib/common' // tslint:disable-line:no-d
 import * as ui from '@kui/core/tests/lib/ui'
 const { cli, selectors, sidecar } = ui
 
-import { expectedVersion } from '../core1/version'
+import { expectedVersion } from '../core/version'
 
 /**
  *
