+++ packages/app/src/test/core/version.ts
