+++ plugins/bash-like/src/test/bash-like/bash-like.ts
