+++ plugins/bash-like/src/test/bash-like/blank-lines.ts
