+++ plugins/bash-like/src/test/bash-like/cd.ts
