+++ plugins/bash-like/src/test/bash-like/ctrl-c.ts
