+++ plugins/bash-like/src/test/bash-like/echo.ts
