+++ plugins/core-support/src/test/core-support/clear.ts
