+++ plugins/core-support/src/test/core-support/screenshot.ts
