+++ plugins/core-support/src/test/core-support/tab-completion.ts
