+++ plugins/core-support/src/test/core-support/text-search.ts
