+++ plugins/field-installed-plugins/src/test/field-installed-plugins/plugin.ts
