+++ plugins/openwhisk/src/test/openwhisk1/command-not-found-suggestions.ts
@@ -21,7 +21,7 @@ import * as ui from '@kui/core/tests/lib/ui'
 import * as openwhisk from '@kui-plugin/openwhisk/tests/lib/openwhisk/openwhisk'
 const { cli, selectors, sidecar } = ui
 
-import { expectSuggestionsFor } from '@kui/core/test/core2/command-not-found-suggestions'
+import { expectSuggestionsFor } from '@kui/core/test/core/command-not-found-suggestions'
 
 describe('Suggestions for command not found', function (this: common.ISuite) {
   before(openwhisk.before(this))
